+++ include/httpRequest.hpp
@@ -6,7 +6,7 @@
 /*   By: jvorstma <jvorstma@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2024/06/01 17:54:54 by jvorstma      #+#    #+#                 */
-/*   Updated: 2024/08/05 18:27:39 by jvorstma      ########   odam.nl         */
+/*   Updated: 2024/08/08 10:35:32 by jvorstma      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -56,6 +56,7 @@ class Request {
 		int 				_maxLengthUploadContent;
 		int					_bytesCopied;
 		string				_uploadedFile;
+		bool				_parseFlag;
 	//	ServerBlock			_server;
 	public:
 		Request();
@@ -71,6 +72,7 @@ class Request {
 		void	addHeader(string const key, string const value);
 		void	setHeader(map<string, string> const header);
 		void	setStatusCode(int const statusCode);
+		void	setParseFlag(bool const parseFlag);
 
 		string				getMethod() const;
 		string				getPath() const;
@@ -79,6 +81,7 @@ class Request {
 		int					getStatusCode() const;
 		map<string, string>	getHeaders() const;
 		string				getHeaderValue(const string& key) const;
+		bool				getParseFlag() const;
 		// -----------------------
 		void				setUploadeFile(string uploadedFile);
 		void				setBytesCopied(long bytesCopied);
@@ -96,7 +99,7 @@ class Request {
 };
 
 void	createRequest(vector<char> requestData, Request& request);
-void	checkBody(vector<char> requestData, Request& request);
+void	checkParseReady(vector<char> requestData, Request& request);
 void	handleRequest(const int clientSocket, Request& request, ServerBlock serverBlock);
 
 //https://www.ibm.com/docs/en/app-connect/11.0.0?topic=messages-http-headers
